+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### Bug Fixes
 
-* [ObjectServer] ProGuard was not configered correctly when working with Subscriptions for Query-based Realms.
+* [ObjectServer] ProGuard was not configured correctly when working with Subscriptions for Query-based Realms.
 
 
 ## 5.4.2 (2018-08-09)